backend/requirements.txt
fastapi==0.112.1
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.8.2
python-dotenv==1.0.1
